+++ src/tools/web/client/package.json
@@ -3,6 +3,21 @@
   "version": "1.0.0",
   "description": "",
   "main": "src/index.js",
+  "scripts": {
+    "start": "npm run web",
+    "web": "webpack-dashboard -c green -- webpack-dev-server",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "material-ui": "^0.15.4",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "react-redux": "^4.4.5",
+    "react-tap-event-plugin": "^1.0.0",
+    "redux": "^3.6.0",
+    "redux-promises": "^1.0.0",
+    "rgx": "^0.2.4"
+  },
   "devDependencies": {
     "babel-core": "^6.14.0",
     "babel-loader": "^6.2.5",
@@ -15,21 +30,5 @@
     "webpack": "^1.13.2",
     "webpack-dashboard": "^0.1.8",
     "webpack-dev-server": "^1.15.1"
-  },
-  "scripts": {
-    "start": "npm run web",
-    "web": "webpack-dashboard -c green -- webpack-dev-server",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "dependencies": {
-    "material-ui": "^0.15.4",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-redux": "^4.4.5",
-    "react-tap-event-plugin": "^1.0.0",
-    "react-treeview": "^0.4.5",
-    "redux": "^3.6.0",
-    "redux-promises": "^1.0.0",
-    "rgx": "^0.2.4"
   }
 }
